Added logic for new page event dialog
