+++ src/components/content-editor/event-trigger-editor.tsx
@@ -6,10 +6,9 @@ import { ReduxActions, ReduxState } from "../../store";
 import { setSelectedExhibition } from "../../actions/exhibitions";
 
 import styles from "../../styles/exhibition-view";
-import { WithStyles, withStyles, MenuItem, Select, TextField, Typography, List, ListItem, ListItemSecondaryAction, IconButton, FormControl, InputLabel, Divider, Paper, Box, ListItemText, Accordion, AccordionSummary, AccordionDetails, Dialog, DialogContent, DialogTitle, DialogActions, Button } from "@material-ui/core";
+import { WithStyles, withStyles, MenuItem, Select, TextField, Typography, List, ListItem, ListItemSecondaryAction, IconButton, FormControl, InputLabel, Divider, Paper, Box, ListItemText, Accordion, AccordionSummary, AccordionDetails } from "@material-ui/core";
 import { KeycloakInstance } from "keycloak-js";
-// eslint-disable-next-line max-len
-import { Exhibition, ExhibitionPage, ExhibitionPageEventTrigger, ExhibitionPageEventActionType, ExhibitionPageEventPropertyType, ExhibitionPageEvent, ExhibitionPageEventProperty, PageLayoutView, PageLayoutWidgetType, VisitorVariable, VisitorVariableType } from "../../generated/client";
+import { Exhibition, ExhibitionPage, ExhibitionPageEventTrigger, ExhibitionPageEvent, PageLayoutView, VisitorVariable } from "../../generated/client";
 import { PhysicalButton, PhysicalButtonData } from '../../types';
 import strings from "../../localization/strings";
 import "codemirror/lib/codemirror.css";
@@ -24,7 +23,7 @@ import DeleteIcon from "@material-ui/icons/Delete";
 import AddIcon from "@material-ui/icons/Add";
 import HelpDialog from "../generic/help-dialog";
 import ExpandMoreIcon from "@material-ui/icons/ChevronRight";
-import GenericDialog from "../generic/generic-dialog";
+import PageEventDialog from "./page-event-dialog";
 
 /**
  * Component props
@@ -46,6 +45,7 @@ interface State {
   loading: boolean;
   selectedPageEventIndex?: number;
   showPageEventDialog: boolean;
+  addingNewPageEvent: boolean;
 }
 
 /**
@@ -62,16 +62,17 @@ class EventTriggerEditor extends React.Component<Props, State> {
     super(props);
     this.state = {
       loading: false,
-      showPageEventDialog: false
+      showPageEventDialog: false,
+      addingNewPageEvent: false
     };
   }
 
   /**
    * Component render method
    */
-  public render() {
+  public render = () => {
 
-    return(
+    return (
       <div style={{ marginTop: theme.spacing(2) }}>
         { this.renderTriggerName() }
         <Paper variant="outlined" square style={{ marginTop: theme.spacing(2) }} >
@@ -88,7 +89,7 @@ class EventTriggerEditor extends React.Component<Props, State> {
             <IconButton
               title={ strings.contentEditor.editor.eventTriggers.addEvent }
               color="primary"
-              onClick={ () => this.setState({ showPageEventDialog: true }) }
+              onClick={ this.onAddPageEventClick }
             >
               <AddIcon />
             </IconButton>
@@ -355,377 +356,24 @@ class EventTriggerEditor extends React.Component<Props, State> {
    * Render event options
    */
   private renderEventOptions = () => {
-    const { selectedEventTrigger } = this.props;
-    const { selectedPageEventIndex, showPageEventDialog } = this.state;
+    const { selectedEventTrigger, visitorVariables, pages, view, availableLanguages } = this.props;
+    const { showPageEventDialog, addingNewPageEvent, selectedPageEventIndex } = this.state;
 
-    console.log(showPageEventDialog);
-    console.log(selectedEventTrigger);
-    console.log(selectedPageEventIndex);
-
-    if (!selectedEventTrigger.events || !selectedEventTrigger.events.length) {
-      return null;
-    }
+    const selectedEvent = selectedPageEventIndex === undefined ? undefined : selectedEventTrigger.events![selectedPageEventIndex];
 
     return (
-      <GenericDialog
-        fullWidth
-        title={ strings.contentEditor.editor.eventTriggers.options }
-        open={ showPageEventDialog }
-        cancelButtonText={ strings.generic.cancel }
-        positiveButtonText={ strings.generic.save }
-        onCancel={ () => this.setState({ showPageEventDialog: false }) }
-        onConfirm={ this.onAddPageEventClick }
-        onClose={ () => this.setState({ showPageEventDialog: false }) }
-        error={ false }
-      >
-        <Typography>
-          { strings.contentEditor.editor.eventTriggers.optionsInstructions }
-        </Typography>
-        { this.renderEventActionTypeSelect() }
-        { this.renderEventActionSettings() }
-      </GenericDialog>
-    );
-
-  }
-
-  /**
-   * Render event action type list
-   */
-  private renderEventActionTypeSelect = () => {
-    const selectedActionType = this.getSelectedEventActionType();
-    const eventActionTypeList = Object.values(ExhibitionPageEventActionType).map((actionType, index) => {
-
-      /**
-       * TODO: Remove these when all page event action types have been implemented
-       */
-      if (
-        actionType === ExhibitionPageEventActionType.Setuservalue ||
-        actionType === ExhibitionPageEventActionType.Navigate ||
-        actionType === ExhibitionPageEventActionType.ExecuteWebScript ||
-        actionType === ExhibitionPageEventActionType.StartVisitorSession
-      ) {
-        return (
-          <MenuItem key={ `eventActionType-${ index }` } value={ actionType }>
-            { strings.contentEditor.editor.eventTriggers.actionTypes[actionType] }
-          </MenuItem>
-        );
-      }
-
-      return null;
-    });
-
-    return (
-      <Box mt={ 2 }>
-        <Select
-          value={ selectedActionType || "" }
-          onChange={ this.onSelectEventActionType }
-        >
-          { eventActionTypeList }
-        </Select>
-      </Box>
-    );
-  }
-
-  /**
-   * Render event action settings based on the selected event action type
-   */
-  private renderEventActionSettings = () => {
-    switch (this.getSelectedEventActionType()) {
-      case ExhibitionPageEventActionType.Setuservalue:
-        return this.renderSetUserValueSettings();
-      case ExhibitionPageEventActionType.Navigate:
-        return this.renderNavigateSettings();
-      case ExhibitionPageEventActionType.Triggerdevicegroupevent:
-        return this.renderDeviceGroupEventSettings();
-      case ExhibitionPageEventActionType.ExecuteWebScript:
-        return this.renderExecuteWebScriptSettings();
-      case ExhibitionPageEventActionType.StartVisitorSession:
-        return this.renderStartVisitorSessionSettings();
-      /**
-       * TODO: Needs implementation
-       */
-      case ExhibitionPageEventActionType.Hide:
-      case ExhibitionPageEventActionType.Show:
-      case ExhibitionPageEventActionType.Setsrc:
-      case ExhibitionPageEventActionType.Settext:
-        return <h5>{ strings.comingSoon }</h5>;
-      default:
-        return <></>;
-    }
-  }
-
-  /**
-   * Render set user value settings
-   */
-  private renderSetUserValueSettings = () => {
-    const { classes, visitorVariables } = this.props;
-
-    const actionType = this.getSelectedEventActionType();
-    if (actionType !== ExhibitionPageEventActionType.Setuservalue) {
-      return;
-    }
-
-    const event = this.getCurrentPageEvent();
-    const userValuePropertyName = event ? event.properties.find(property => property.name === "name") : undefined;
-    const userValuePropertyValue = event ? event.properties.find(property => property.name === "value") : undefined;
-    const variableName = userValuePropertyName?.value || "";
-    const variableValue = userValuePropertyValue?.value || "";
-
-    return (
-      <div style={{ marginTop: theme.spacing(2) }}>
-        <Typography variant="h6">
-          { strings.contentEditor.editor.eventTriggers.variableName }
-        </Typography>
-
-        <Select 
-          fullWidth={ false }
-          name="name"
-          value={ variableName }
-          onChange={ this.onEventTriggerEventPropertyChange }
-          className={ classes.selectResourceEditor } >
-          { 
-            visitorVariables.map(visitorVariable => {
-              return <MenuItem value={ visitorVariable.name }>{ visitorVariable.name }</MenuItem>
-            }) 
-          }
-        </Select>
-
-        <Typography variant="h6">
-          { strings.contentEditor.editor.eventTriggers.variableValue }
-        </Typography>
-
-        { this.renderSetUserValueSettingValueInput(variableName, variableValue) }
-      </div>
-    );
-  }
-
-  /**
-   * Renders input for set user variable trigger
-   * 
-   * @param variableName variable name
-   * @param variableValue variable value
-   */
-  private renderSetUserValueSettingValueInput = (variableName: string, variableValue: string) => {
-    const { classes, visitorVariables } = this.props;
-
-    const visitorVariable = visitorVariables.find(item => item.name === variableName); 
-    const visitorVariableType = visitorVariable?.type || VisitorVariableType.Text;
-
-    if (visitorVariableType === VisitorVariableType.Enumerated) {
-      const values = visitorVariable?._enum || [];
-      return (
-        <Select 
-          fullWidth={ false }
-          name="value"
-          value={ variableValue }
-          onChange={ this.onEventTriggerEventPropertyChange }
-          className={ classes.selectResourceEditor } >
-          { 
-            values.map(value => {
-              return <MenuItem value={ value }>{ value }</MenuItem>
-            }) 
-          }
-        </Select>
-      );
-    }
-
-    if (visitorVariableType === VisitorVariableType.Boolean) {
-      return (
-        <Select 
-          fullWidth={ false }
-          name="value"
-          value={ variableValue }
-          onChange={ this.onEventTriggerEventPropertyChange }
-          className={ classes.selectResourceEditor } 
-        >
-          <MenuItem value={ "true" }>{ strings.contentEditor.editor.eventTriggers.variableBooleanTrue }</MenuItem>
-          <MenuItem value={ "false" }>{ strings.contentEditor.editor.eventTriggers.variableBooleanFalse }</MenuItem>
-        </Select>
-      );
-    }
-    
-    return (
-      <TextField
-        fullWidth={ false }
-        name="value"
-        type={ visitorVariableType === VisitorVariableType.Number ? "number" : "text" }
-        className={ classes.textResourceEditor }
-        value={ variableValue }
-        onChange={ this.onEventTriggerEventPropertyChange }
+      <PageEventDialog
+        selectedEvent={ selectedEvent }
+        pages={ pages }
+        visitorVariables={ visitorVariables }
+        creatingNew={ addingNewPageEvent }
+        dialogOpen={ showPageEventDialog }
+        view={ view }
+        availableLanguages={ availableLanguages }
+        onCloseOrCancel={ () => this.setState({ showPageEventDialog: false }) }
+        onCreate={ this.onAddNewPageEventClick }
+        onUpdate={ this.updateEventInJson }
       />
-    );
-  }
-
-  /**
-   * Render navigation settings
-   */
-  private renderNavigateSettings = () => {
-    if (this.getSelectedEventActionType() !== ExhibitionPageEventActionType.Navigate) {
-      return;
-    }
-
-    const selectedPageEvent = this.getCurrentPageEvent();
-
-    if (!selectedPageEvent) {
-      return null;
-    }
-
-    const property = selectedPageEvent.properties.find(prop => prop.name === "pageId");
-    return (
-      <div style={{ marginTop: theme.spacing(2) }}>
-        <Typography variant="h6">
-          { strings.contentEditor.editor.eventTriggers.selectPage }
-        </Typography>
-        <Select
-          name={ "pageId" }
-          value={ property ? property.value : "" }
-          onChange={ this.onEventTriggerEventPropertyChange }
-        >
-          { this.fetchPagesInExhibition() }
-        </Select>
-      </div>
-    );
-  }
-
-  /**
-   * Render device group event settings
-   */
-  private renderDeviceGroupEventSettings = () => {
-    const { selectedEventTrigger, classes } = this.props;
-    const actionType = this.getSelectedEventActionType();
-    if (actionType !== ExhibitionPageEventActionType.Triggerdevicegroupevent) {
-      return;
-    }
-
-    const event = (selectedEventTrigger.events && selectedEventTrigger.events.length) ? selectedEventTrigger.events[0] : undefined;
-    const deviceGroupEventNameProperty = event ? event.properties.find(property => property.name === "name") : undefined;
-
-    return (
-      <div style={{ marginTop: theme.spacing(2) }}>
-        <Typography variant="h6">
-          { strings.contentEditor.editor.eventTriggers.deviceGroupEvent }
-        </Typography>
-        <TextField
-          fullWidth={ false }
-          name="name"
-          className={ classes.textResourceEditor }
-          value={ deviceGroupEventNameProperty?.value || "" }
-          onChange={ this.onEventTriggerEventPropertyChange }
-        />
-      </div>
-    );
-  }
-
-  /**
-   * Renders start visitor session action settings
-   */
-  private renderStartVisitorSessionSettings = () => {
-    if (this.getSelectedEventActionType() !== ExhibitionPageEventActionType.StartVisitorSession) {
-      return;
-    }
-
-    const selectedPageEvent = this.getCurrentPageEvent();
-    if (!selectedPageEvent) {
-      return;
-    }
-    
-    const language = selectedPageEvent.properties.find(prop => prop.name === "language")?.value;
-
-    return (
-      <div style={{ marginTop: theme.spacing(2) }}>
-        <Typography variant="h6">
-          { strings.contentEditor.editor.eventTriggers.selectLanguage }
-        </Typography>
-        <Select
-          name={ "language" }
-          value={ language }
-          onChange={ this.onEventTriggerEventPropertyChange }
-        >
-          { this.renderLanguageOptions() }
-        </Select>
-      </div>
-    );
-  }
-  /**
-   * Renders language options
-   * 
-   * @returns options as JSX Elements or undefined if no options found
-   */
-  private renderLanguageOptions = (): JSX.Element[] | undefined => {
-    const { availableLanguages } = this.props;
-
-    return availableLanguages.map(language =>
-      <MenuItem key={ language } value={ language }>
-        { language }
-      </MenuItem>
-    );
-  }
-
-  /**
-   * Renders execute web script settings
-   */
-  private renderExecuteWebScriptSettings = () => {
-    const { classes, view } = this.props;
-    if (this.getSelectedEventActionType() !== ExhibitionPageEventActionType.ExecuteWebScript) {
-      return;
-    }
-
-    const selectedPageEvent = this.getCurrentPageEvent();
-    if (!selectedPageEvent) {
-      return;
-    }
-
-    const webViewOptions = this.getWebViewsInLayout(view);
-    if (webViewOptions.length < 1) {
-      return (
-        <div style={{ marginTop: theme.spacing(2) }}>
-          <Typography>
-            { strings.contentEditor.editor.eventTriggers.noWebViews }
-          </Typography>
-        </div>
-      );
-    }
-
-    const webViewProperty = selectedPageEvent.properties.find(prop => prop.name === "webViewId");
-    const scriptProperty = selectedPageEvent.properties.find(prop => prop.name === "script");
-
-    return (
-      <>
-        <div style={{ marginTop: theme.spacing(2) }}>
-          <Typography variant="h6">
-            { strings.contentEditor.editor.eventTriggers.selectPage }
-          </Typography>
-          <Select
-            name={ "webViewId" }
-            value={ webViewProperty?.value ?? "" }
-            onChange={ this.onEventTriggerEventPropertyChange }
-          >
-            { this.renderWebViewOptions(webViewOptions) }
-          </Select>
-        </div>
-        <div style={{ marginTop: theme.spacing(2) }}>
-          <TextField
-            fullWidth={ false }
-            name="script"
-            className={ classes.textResourceEditor }
-            value={ scriptProperty?.value ?? "" }
-            onChange={ this.onEventTriggerEventPropertyChange }
-          />
-        </div>
-      </>
-    );
-  }
-
-  /**
-   * Renders web view options
-   * 
-   * @returns options as JSX Elements or undefined if no options found
-   */
-  private renderWebViewOptions = (webViewOptions: PageLayoutView[]): JSX.Element[] | undefined => {
-    return webViewOptions.map(view =>
-      <MenuItem key={ view.id } value={ view.id }>
-        { view.name ?? view.id }
-      </MenuItem>
     );
   }
 
@@ -750,78 +398,6 @@ class EventTriggerEditor extends React.Component<Props, State> {
     });
 
     return items;
-  }
-
-  /**
-   * Gets action type of selected event
-   *
-   * @returns exhibition page event action type of undefined
-   */
-  private getSelectedEventActionType = (): ExhibitionPageEventActionType | undefined => {
-    const { selectedEventTrigger } = this.props;
-    const { selectedPageEventIndex } = this.state;
-    if (selectedPageEventIndex === undefined) {
-      return;
-    }
-
-    const event = selectedEventTrigger.events && selectedEventTrigger.events.length ? 
-      selectedEventTrigger.events[selectedPageEventIndex] :
-      undefined;
-
-    return event ? event.action : undefined;
-  }
-
-  /**
-   * Get current page event
-   *
-   * @returns found exhibition page event or undefined
-   */
-  private getCurrentPageEvent = (): ExhibitionPageEvent | undefined => {
-    const { selectedEventTrigger } = this.props;
-    const { selectedPageEventIndex } = this.state;
-
-    if (selectedPageEventIndex === undefined || !selectedEventTrigger.events) {
-      return;
-    }
-
-    return selectedEventTrigger.events[selectedPageEventIndex];
-  }
-
-  /**
-   * Render pages in exhibition
-   */
-  private fetchPagesInExhibition = () => {
-    const pageSelectionItems = this.props.pages.map(page => {
-      return (
-        <MenuItem key={ `event-trigger-navigation-${page.id}` } value={ page.id }>
-          { page.name }
-        </MenuItem>
-      );
-    });
-    return pageSelectionItems;
-  }
-
-  /**
-   * Get web views in page layout
-   * 
-   * @param view page layout view
-   * @returns web views as page layout views
-   */
-  private getWebViewsInLayout = (view?: PageLayoutView): PageLayoutView[] => {
-    const webViews: PageLayoutView[] = [];
-    if (view) {
-      if (view.widget === PageLayoutWidgetType.WebView) {
-        webViews.push(view);
-      }
-  
-      if (view.children) {
-        view.children.forEach(childView => {
-          webViews.push(...this.getWebViewsInLayout(childView));
-        });
-      }
-    }
-
-    return webViews;
   }
 
   /**
@@ -874,14 +450,23 @@ class EventTriggerEditor extends React.Component<Props, State> {
   }
 
   /**
-   * Event handler for add page event click
+   * Event handler for add new page event click
    */
   private onAddPageEventClick = () => {
+    this.setState({
+      showPageEventDialog: true,
+      addingNewPageEvent: true,
+      selectedPageEventIndex: undefined
+    });
+  }
+
+  /**
+   * Event handler for add page event click
+   *
+   * @param newPageEvent new page event to be saved
+   */
+  private onAddNewPageEventClick = (newPageEvent: ExhibitionPageEvent) => {
     const { onSave } = this.props;
-    const newPageEvent: ExhibitionPageEvent = {
-      action: ExhibitionPageEventActionType.Navigate,
-      properties: []
-    };
 
     const trigger: ExhibitionPageEventTrigger = { ...this.props.selectedEventTrigger };
     if (!trigger.events) {
@@ -891,7 +476,9 @@ class EventTriggerEditor extends React.Component<Props, State> {
     trigger.events = [ ...trigger.events, newPageEvent ];
 
     this.setState({
-      selectedPageEventIndex: trigger.events ? trigger.events.length - 1 : undefined
+      selectedPageEventIndex: trigger.events ? trigger.events.length - 1 : undefined,
+      showPageEventDialog: false,
+      addingNewPageEvent: false,
     });
 
     onSave(trigger);
@@ -905,7 +492,8 @@ class EventTriggerEditor extends React.Component<Props, State> {
   private onPageEventClick = (pageEventIndex: number) => () => {
     this.setState({
       selectedPageEventIndex: pageEventIndex,
-      showPageEventDialog: true
+      showPageEventDialog: true,
+      addingNewPageEvent: false
     });
   }
 
@@ -930,139 +518,34 @@ class EventTriggerEditor extends React.Component<Props, State> {
   }
 
   /**
-   * Event handler for event trigger event property change
+   * Event handler for updating page event in json
    *
-   * @param event react change event
+   * @param updatedPageEvent updated page event
    */
-  private onEventTriggerEventPropertyChange = (event: React.ChangeEvent<HTMLInputElement | { name?: string; value: any }>) => {
-    const { onSave } = this.props;
-    const { selectedPageEventIndex } = this.state;
-
-    const trigger: ExhibitionPageEventTrigger = { ...this.props.selectedEventTrigger };
-    const propertyName = event.target.name;
-    const propertyValue = event.target.value as string;
-    if (propertyName === undefined || !trigger.events || selectedPageEventIndex === undefined) {
-      return;
-    }
-
-    trigger.events = produce(trigger.events, draft => {
-      const propertyIndex = draft[selectedPageEventIndex].properties.findIndex(property => property.name === propertyName);
-      if (propertyIndex === -1) {
-        draft[selectedPageEventIndex].properties.push({
-          name: propertyName,
-          value: propertyValue,
-          type: ExhibitionPageEventPropertyType.String
-        });
-      } else {
-        draft[selectedPageEventIndex].properties[propertyIndex].value = propertyValue;
-      }
-    });
-
-    onSave(trigger);
-  }
-
-  /**
-   * On select event action handler
-   *
-   * @param event react change event
-   */
-  private onSelectEventActionType = (event: React.ChangeEvent<{ value: unknown }>) => {
-    const value = event.target.value as ExhibitionPageEventActionType;
-
-    /**
-     * TODO: If we assume that each event trigger contains only one event, do we want to clear the existing values on select?
-     */
-    this.overwriteEventInJson(value);
-  }
-
-  /**
-   * Overwrite event in json handler
-   *
-   * @param actionType new event action type
-   */
-  private overwriteEventInJson = (actionType: ExhibitionPageEventActionType) => {
+  private updateEventInJson = (updatedPageEvent: ExhibitionPageEvent) => {
     const { onSave } = this.props;
     const { selectedPageEventIndex } = this.state;
     const trigger = { ...this.props.selectedEventTrigger } as ExhibitionPageEventTrigger;
+  
+    if (selectedPageEventIndex === undefined) {
+      return;
+    }
 
-    const newAction = this.createEvent(actionType);
-
-    if (!trigger.events || !selectedPageEventIndex) {
+    if (!trigger.events) {
       trigger.events = [];
-      trigger.events.push(newAction);
+      trigger.events.push(updatedPageEvent);
     } else {
       trigger.events = produce(trigger.events, draft => {
-        draft.splice(selectedPageEventIndex, 1, newAction);
+        draft.splice(selectedPageEventIndex, 1, updatedPageEvent);
       });
     }
 
+    this.setState({
+      showPageEventDialog: false,
+      addingNewPageEvent: false,
+    });
+
     onSave(trigger);
-  }
-
-  /**
-   * Get event properties based on action type
-   *
-   * @param actionType ExhibitionPageEventActionType
-   * @returns list of exhibition page event properties
-   */
-  private getEventPropertiesByType = (actionType: ExhibitionPageEventActionType): ExhibitionPageEventProperty[] => {
-    switch (actionType) {
-      case ExhibitionPageEventActionType.Navigate:
-        return [
-          this.getStringProperty("pageId")
-        ];
-      case ExhibitionPageEventActionType.Setuservalue:
-        return [
-          this.getStringProperty("name"),
-          this.getStringProperty("value")
-        ];
-      case ExhibitionPageEventActionType.Triggerdevicegroupevent:
-        return [
-          this.getStringProperty("name")
-        ];
-      case ExhibitionPageEventActionType.ExecuteWebScript:
-        return [
-          this.getStringProperty("webViewId"),
-          this.getStringProperty("script")
-        ];
-      case ExhibitionPageEventActionType.Hide:
-      case ExhibitionPageEventActionType.Show:
-      case ExhibitionPageEventActionType.Setsrc:
-      case ExhibitionPageEventActionType.Settext:
-      default:
-        return [];
-    }
-  }
-
-  /**
-   * Get string type property
-   *
-   * @param name property name
-   * @param value property value
-   * @returns exhibition page event property
-   */
-  private getStringProperty = (name: string, value?: string): ExhibitionPageEventProperty => {
-    return {
-      name: name,
-      value: value || "",
-      type: ExhibitionPageEventPropertyType.String
-    };
-  }
-
-  /**
-   * Create event based on the given values
-   *
-   * @param actionType given action type
-   * @returns exhibition page event
-   */
-  private createEvent = (actionType: ExhibitionPageEventActionType): ExhibitionPageEvent => {
-    const eventProperties: ExhibitionPageEventProperty[] = this.getEventPropertiesByType(actionType);
-    const newEvent: ExhibitionPageEvent = {
-      action: actionType,
-      properties: eventProperties
-    };
-
-    return newEvent;
   }
 }
